backend/requirements-prod.txt
fastapi
uvicorn
pypdf
pinecone
groq
google-genai
mlflow
python-dotenv
python-multipart
numpy<2
torch==2.4.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers==2.7.0
transformers==4.41.2
python-jose[cryptography]
passlib[bcrypt]